Moved report helpers to component
